simplify TestCommand.run
